Omit CLI args on local runs
